// apps/web/src/shared/api/gen/model/listOrders200DataItemPaymentMethod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Checkout API
 * POST /api/sessions с телом {} создаёт гостевую сессию. Вставьте data.token в Authorize без слова Bearer. Суммы указаны в копейках. Данные сохраняются после перезапуска. Оплата тестовая.
 * OpenAPI spec version: 1.0.0
 */

export const ListOrders200DataItemPaymentMethod = {
  card: 'card',
  cash_on_delivery: 'cash_on_delivery',
} as const;